(PC-16449) feat: remove FORCE_PHONE_VALIDATION flag
